Parallax Website
Created a Parallax Website using:

HTML
CSS
Parallax is a computer graphics technique used by web designers to create a faux-3D effect, it is a technique in which the website background moves at a different speed than the foreground.

I have used a property of CSS called background-attachment. *This property is used to set whether a background image is fixed or scrollable.